Move RSK initialization outside Bean creation scope
